+++ src/store/storable.ts
@@ -1,26 +1,33 @@
 import Writable from './store.js';
-import serialize from 'serialize-javascript';
-import { deserialize } from '../lib/helpers.js';
 
-class Storable extends Writable {
+export class Storable extends Writable {
     private key
     private detached = false
 
     constructor(key : string, value? : any, start? : Function) {
         super(value, start);
         this.key = key;
-        this.retrieve();
+        this.retrieve(value);
         this.detach = this.detach.bind(this);
         this.attach = this.attach.bind(this);
     }
 
-    private save(value : any):void {
-        if (!this.detached) localStorage.setItem(this.key, serialize(value));
+    protected serialize(value) {
+        return JSON.stringify(value)
     }
 
-    private retrieve():void {
-        const value = localStorage.getItem(this.key);
-        if (!!value) this.set(deserialize(value));
+    protected deserialize(data) {
+        return JSON.parse(data)
+    }
+
+    private save(value : any):void {
+        if (!this.detached) localStorage.setItem(this.key, this.serialize(value));
+    }
+
+    private retrieve(value : any):void {
+        const saved = localStorage.getItem(this.key);
+        if (!!saved) this.set(this.deserialize(saved));
+        else this.save(value)
     }
 
     set(newValue : any):void {