+++ test/browser/src/browser.test.js
@@ -1,3 +1,5 @@
 import storable from './store/storable.test.js'
+import serializable from './store/serializable.test.js'
 
 storable()
+serializable()