+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # 2.1.0: November 30, 2020
-- Added the ability to serialize and deserialize javascript apis like `Map` and `() => doSomething()`
+- Added the serializable store to be able to store javascript apis like `Map` and `() => doSomething()`